requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# ip_webcam.py: 3
# main_script.py: 23
Pillow == 4.1.0

# main_script.py: 22
matplotlib == 2.0.2

# main_script.py: 12
numpy == 1.13.1

# ip_webcam.py: 1
requests == 2.13.0

# main_script.py: 14
six == 1.10.0

# main_script.py: 44,46
utils == 0.9.0
